requirements.txt
streamlit
opencv-python-headless
ultralytics
numpy
stqdm
deep_sort_realtime
python-dotenv
twilio
